Move sessions as a telescope extension
